Fix: refactor unused, misplaced, and missing stubs
